refactor(#800): split providers/copilot_chat.rs by moving tests to sibling
Move 4,335-line providers/copilot_chat.rs into the existing
providers/copilot_chat/ directory as mod.rs, and extract the 2,190-line
inline tests block into a sibling tests.rs (matching the dir-as-module
pattern already used for jetbrains.rs).

- copilot_chat/mod.rs:   2,147 LOC (production code)
- copilot_chat/tests.rs: 2,188 LOC (unit tests)
- copilot_chat/jetbrains.rs: 2,042 LOC (unchanged)

Production module now well below 2,500 LOC. Test fixtures re-rooted
(include_str! paths drop the `copilot_chat/` prefix since tests.rs now
lives alongside the fixtures/ directory). `super::*` continues to grant
the test module access to private items.

`cargo test -p budi-core --lib` green.
